guest user cart functionality

+++ phoenixsite/productManagement/models.py
@@ -6,9 +6,9 @@ class Product(models.Model):
     product_id = models.IntegerField(primary_key=True)
     product_name = models.CharField(max_length = 350) 
     product_image = models.ImageField(null=True, blank=True)
-    product_price = models.FloatField() #sale price per product
+    product_price = models.DecimalField(max_digits=10, decimal_places=2) #sale price per product
     digital = models.BooleanField(default=False, null=True, blank=True) #ship product if its not digital, hence the default false
-    product_tax = models.FloatField()
+    product_tax = models.DecimalField(max_digits=10, decimal_places=2)
     stock = models.IntegerField() #automate? goods not sold/ in storage
     products_bought = models.IntegerField() #no of products bought,
     purchase_price = models.IntegerField()# amount used to purchase the products for sale
@@ -32,7 +32,7 @@ class Order (models.Model): ##This basically represents the CART
     order_id = models.AutoField(primary_key=True, unique=True, auto_created=True)
     order_date = models.DateField(auto_now=True) 
     complete = models.BooleanField(default=False) #if complete is false can continue adding items, changes status of cart is it same as order status??
-    cost = models.FloatField(null=True, blank=True) # total quantity * product_price - discount 
+    cost = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True) # total quantity * product_price - discount 
     def __str__(self):
         return str(self.order_id)
     
@@ -89,7 +89,7 @@ class ShippingAddress(models.Model):
     country =models.CharField(max_length=250, null=True, blank=True)
     city = models.CharField(max_length=200)
     zipcode = models.CharField(max_length=200)
-    delivery_cost = models.FloatField() #depending with location, product_weight, any discounts
+    delivery_cost = models.DecimalField(max_digits=10, decimal_places=2) #depending with location, product_weight, any discounts
     shipping_date = models.DateTimeField(auto_now_add=True) 
     def __str__(self):
         return self.shipping_code
@@ -115,7 +115,7 @@ class PaymentInfo(models.Model):
     payment_code = models.CharField(max_length= 20, primary_key=True)
     payment_type = models.CharField(max_length= 250, help_text="Mpesa(Kenyan phone numbers only)")
     payment_phone = models.IntegerField(help_text='0712345678 or +254712345678') #look for a validator, ie. regex 
-    amount = models.FloatField() #amount being paid
+    amount = models.DecimalField(max_digits=10, decimal_places=2) #amount being paid
     payment_status = models.CharField(max_length= 250)
     def __str__(self):
         return self.payment_code
@@ -127,7 +127,7 @@ class Offer(models.Model): #can I apply these discounts to the other tables and 
     shippingAddress = models.ForeignKey(ShippingAddress, models.SET_NULL, blank=True, null = True)
 
     coupon_code = models.CharField(max_length = 20, primary_key=True)
-    offer_discount = models.FloatField() #Must be in percentages
+    offer_discount = models.DecimalField(max_digits=10, decimal_places=2) #Must be in percentages
     description = models.CharField(max_length=1000)
     def __str__(self):
         return self.coupon_code
